Organizing Typescript API.
