+++ src/zts-helpers.ts
@@ -3,8 +3,6 @@ import { z } from "zod";
 import { FlatObject } from "./common-helpers";
 import { SchemaHandler } from "./schema-walker";
 
-const { factory: f } = ts;
-
 export type LiteralType = string | number | boolean;
 
 export interface ZTSContext extends FlatObject {
@@ -17,38 +15,6 @@ export interface ZTSContext extends FlatObject {
 }
 
 export type Producer = SchemaHandler<ts.TypeNode, ZTSContext>;
-
-export const addJsDocComment = <T extends ts.Node>(node: T, text: string) =>
-  ts.addSyntheticLeadingComment(
-    node,
-    ts.SyntaxKind.MultiLineCommentTrivia,
-    `* ${text} `,
-    true,
-  );
-
-export const printNode = (
-  node: ts.Node,
-  printerOptions?: ts.PrinterOptions,
-) => {
-  const sourceFile = ts.createSourceFile(
-    "print.ts",
-    "",
-    ts.ScriptTarget.Latest,
-    false,
-    ts.ScriptKind.TS,
-  );
-  const printer = ts.createPrinter(printerOptions);
-  return printer.printNode(ts.EmitHint.Unspecified, node, sourceFile);
-};
-
-const safePropRegex = /^[A-Za-z_$][A-Za-z0-9_$]*$/;
-
-export const makePropertyIdentifier = (name: string | number) =>
-  typeof name === "number"
-    ? f.createNumericLiteral(name)
-    : safePropRegex.test(name)
-      ? f.createIdentifier(name)
-      : f.createStringLiteral(name);
 
 const primitives: ts.KeywordTypeSyntaxKind[] = [
   ts.SyntaxKind.AnyKeyword,
