+++ tests/unit/zts.spec.ts
@@ -1,9 +1,9 @@
 import ts from "typescript";
 import { z } from "zod";
 import { ez } from "../../src";
-import { f } from "../../src/integration-helpers";
+import { f, printNode } from "../../src/typescript-api";
 import { zodToTs } from "../../src/zts";
-import { ZTSContext, printNode } from "../../src/zts-helpers";
+import { ZTSContext } from "../../src/zts-helpers";
 
 describe("zod-to-ts", () => {
   const printNodeTest = (node: ts.Node) =>
